fixed merge conflicts so all our code can live happily together
